Refactor: excpetion 패키지 생성
